Simplify source code structure and remove unnecessary files
